requirements.txt
google-generativeai==0.5.4
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.3
streamlit==1.35.0
python-dotenv==1.0.1
pytest==8.2.1
ruff==0.4.4